feat(tabBar): update to use Tailwind
